refactor: simplify notifications without audio
